User list order

+++ product/deleteproduct.php
@@ -1,5 +1,16 @@
 <?php
 
+    session_start();
+    if (!isset($_SESSION['user_name'])) {
+        header("Location: ../user/login.php");
+        exit();
+    }
+
+    if ( $_SESSION['role'] !== 'admin' ) {
+        header("Location: ../unauthorized.php");
+        exit();
+    }
+    
     include_once '../connect.php';
     $productid = $_GET['productid'];
     $sql = "DELETE FROM products WHERE id = '$productid'";
